moved board component into features
